Revert "Store"
