+++ package.json
@@ -14,7 +14,6 @@
     "react-router": "^5.0.1",
     "react-router-dom": "^5.0.1",
     "react-scripts": "3.0.1",
-    "react-transition-group": "^1.2.1",
     "react-with-firebase-auth": "^1.0.3",
     "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
