+++ src/components/header/index.js
@@ -3,8 +3,9 @@ import { connect } from 'react-redux'
 import styled from 'styled-components'
 import Clock from './Clock'
 import MoodBar from './MoodBar'
-import Age from './Age'
 import get from 'lodash/get'
+import { timeDifference } from '../../utils/time'
+import { getPlural } from '../../utils/string'
 
 const StyledHeader = styled.header`
   display: flex;
@@ -21,6 +22,10 @@ const StyledHeader = styled.header`
     flex-grow: 1;
     align-items: center;
     margin: 0 10px;
+    .Neko_age {
+      letter-spacing: 1px;
+      font-size: 12px;
+    }
   }
   .Neko_infos {
     display: flex;
@@ -38,21 +43,29 @@ const StyledHeader = styled.header`
   }
 `
 
-const Header = ({ neko }) => (
-  <StyledHeader className="App-header">
-    <div className="Neko_infos">
-      <img src={require('../../img/logo_neko_nect.jpg')} className="Logo" alt="logo" />
-      <div className="Neko_name">
-        {get(neko, 'name')}
+const Header = ({ neko }) => {
+  const from = get(neko, 'birthdate').toDate()
+  const now = new Date()
+  const age = timeDifference(from, now)
+  const { days, hours, minutes } = age
+  return (
+    <StyledHeader className="App-header">
+      <div className="Neko_infos">
+        <img src={require('../../img/logo_neko_nect.jpg')} className="Logo" alt="logo" />
+        <div className="Neko_name">
+          {get(neko, 'name')}
+        </div>
       </div>
-    </div>
-    <div className="Neko_infos_life">
-      <MoodBar />
-      <Age birthdate={get(neko, 'birthdate')} />
-    </div>
-    <Clock />
-  </StyledHeader>
-)
+      <div className="Neko_infos_life">
+        <MoodBar age={age} />
+        <div className="Neko_age">
+          {days} jour{getPlural(days, 1, 's')}, {hours} h et {minutes} min
+        </div>
+      </div>
+      <Clock />
+    </StyledHeader>
+  )
+}
 
 const mapStateToProps = state => ({
   neko: get(state, 'neko')
