+++ src/reducers/index.js
@@ -2,15 +2,11 @@ import { combineReducers } from 'redux'
 import mood from './mood'
 import time from './time'
 import neko from './neko'
-import store from './store'
-import inventory from './inventory'
 
 const reducers = combineReducers({
   mood,
   time,
-  neko,
-  store,
-  inventory
+  neko
 })
 
 export default reducers
